alter tables to add deleted column, alter functions to set a value upon creation
